:technologist: Implement clock module which can switch out real system time for a mocked system time
:white_check_mark: Use mocked time to verify store expiry test

:label: Use u64 as consistent type for epoch seconds/millis

:arrow_up: Upgrade tokio crate

+++ src/resp/command.rs
@@ -42,8 +42,8 @@ pub enum SetCommandExistFlag {
 
 pub enum SetCommandTTLFlag {
     EX(u64),    // TTL duration (seconds)
-    PX(u128),   // TTL duration (milliseconds)
+    PX(u64),   // TTL duration (milliseconds)
     EXAT(u64),  // Set expiry at exact unix time (seconds)
-    PXAT(u128), // Set expiry at exact unix time (milliseconds)
+    PXAT(u64), // Set expiry at exact unix time (milliseconds)
     KEEPTTL,    // Keep existing TTL when setting value
 }
